+++ src/ws.cpp
@@ -25,9 +25,6 @@
 
 #include "ws.h"
 
-#include <EGL/egl.h>
-#include <EGL/eglext.h>
-#include "linux-dmabuf/linux-dmabuf.h"
 #include "bridge/wpe-bridge-server-protocol.h"
 #include "video-plane-display-dmabuf/wpe-video-plane-display-dmabuf-server-protocol.h"
 #include <cassert>
@@ -35,51 +32,6 @@
 #include <string.h>
 #include <sys/socket.h>
 #include <unistd.h>
-
-#ifndef EGL_WL_bind_wayland_display
-#define EGL_WAYLAND_BUFFER_WL 0x31D5
-typedef EGLBoolean (EGLAPIENTRYP PFNEGLBINDWAYLANDDISPLAYWL) (EGLDisplay dpy, struct wl_display *display);
-typedef EGLBoolean (EGLAPIENTRYP PFNEGLQUERYWAYLANDBUFFERWL) (EGLDisplay dpy, struct wl_resource *buffer, EGLint attribute, EGLint *value);
-#endif
-
-#ifndef EGL_EXT_image_dma_buf_import_modifiers
-typedef EGLBoolean (* PFNEGLQUERYDMABUFFORMATSEXTPROC) (EGLDisplay dpy, EGLint max_formats, EGLint *formats, EGLint *num_formats);
-typedef EGLBoolean (* PFNEGLQUERYDMABUFMODIFIERSEXTPROC) (EGLDisplay dpy, EGLint format, EGLint max_modifiers, EGLuint64KHR *modifiers, EGLBoolean *external_only, EGLint *num_modifiers);
-#endif /* EGL_EXT_image_dma_buf_import_modifiers */
-
-#ifndef EGL_EXT_image_dma_buf_import
-#define EGL_LINUX_DMA_BUF_EXT             0x3270
-#define EGL_DMA_BUF_PLANE0_FD_EXT         0x3272
-#define EGL_DMA_BUF_PLANE0_OFFSET_EXT     0x3273
-#define EGL_DMA_BUF_PLANE0_PITCH_EXT      0x3274
-#define EGL_DMA_BUF_PLANE1_FD_EXT         0x3275
-#define EGL_DMA_BUF_PLANE1_OFFSET_EXT     0x3276
-#define EGL_DMA_BUF_PLANE1_PITCH_EXT      0x3277
-#define EGL_DMA_BUF_PLANE2_FD_EXT         0x3278
-#define EGL_DMA_BUF_PLANE2_OFFSET_EXT     0x3279
-#define EGL_DMA_BUF_PLANE2_PITCH_EXT      0x327A
-#endif /* EGL_EXT_image_dma_buf_import */
-
-#ifndef EGL_EXT_image_dma_buf_import_modifiers
-#define EGL_DMA_BUF_PLANE3_FD_EXT         0x3440
-#define EGL_DMA_BUF_PLANE3_OFFSET_EXT     0x3441
-#define EGL_DMA_BUF_PLANE3_PITCH_EXT      0x3442
-#define EGL_DMA_BUF_PLANE0_MODIFIER_LO_EXT 0x3443
-#define EGL_DMA_BUF_PLANE0_MODIFIER_HI_EXT 0x3444
-#define EGL_DMA_BUF_PLANE1_MODIFIER_LO_EXT 0x3445
-#define EGL_DMA_BUF_PLANE1_MODIFIER_HI_EXT 0x3446
-#define EGL_DMA_BUF_PLANE2_MODIFIER_LO_EXT 0x3447
-#define EGL_DMA_BUF_PLANE2_MODIFIER_HI_EXT 0x3448
-#define EGL_DMA_BUF_PLANE3_MODIFIER_LO_EXT 0x3449
-#define EGL_DMA_BUF_PLANE3_MODIFIER_HI_EXT 0x344A
-#endif /* EGL_EXT_image_dma_buf_import_modifiers */
-
-static PFNEGLBINDWAYLANDDISPLAYWL s_eglBindWaylandDisplayWL;
-static PFNEGLQUERYWAYLANDBUFFERWL s_eglQueryWaylandBufferWL;
-static PFNEGLCREATEIMAGEKHRPROC s_eglCreateImageKHR;
-static PFNEGLDESTROYIMAGEKHRPROC s_eglDestroyImageKHR;
-static PFNEGLQUERYDMABUFFORMATSEXTPROC s_eglQueryDmaBufFormatsEXT;
-static PFNEGLQUERYDMABUFMODIFIERSEXTPROC s_eglQueryDmaBufModifiersEXT;
 
 struct wpe_video_plane_display_dmabuf_export {
     struct wl_resource* updateResource;
@@ -132,17 +84,6 @@ GSourceFuncs ServerSource::s_sourceFuncs = {
     nullptr, // closure_marshall
 };
 
-struct Surface {
-    uint32_t id { 0 };
-    struct wl_client* client { nullptr };
-
-    ExportableClient* exportableClient { nullptr };
-
-    struct wl_resource* bufferResource { nullptr };
-    const struct linux_dmabuf_buffer* dmabufBuffer { nullptr };
-    struct wl_shm_buffer* shmBuffer { nullptr };
-};
-
 static const struct wl_surface_interface s_surfaceInterface = {
     // destroy
     [](struct wl_client*, struct wl_resource*) { },
@@ -150,13 +91,7 @@ static const struct wl_surface_interface s_surfaceInterface = {
     [](struct wl_client*, struct wl_resource* surfaceResource, struct wl_resource* bufferResource, int32_t, int32_t)
     {
         auto& surface = *static_cast<Surface*>(wl_resource_get_user_data(surfaceResource));
-
-        surface.dmabufBuffer = Instance::singleton().getDmaBufBuffer(bufferResource);
-        surface.shmBuffer = wl_shm_buffer_get(bufferResource);
-
-        if (surface.bufferResource)
-            wl_buffer_send_release(surface.bufferResource);
-        surface.bufferResource = bufferResource;
+        Instance::singleton().impl().surfaceAttach(surface, bufferResource);
     },
     // damage
     [](struct wl_client*, struct wl_resource*, int32_t, int32_t, int32_t, int32_t) { },
@@ -184,18 +119,7 @@ static const struct wl_surface_interface s_surfaceInterface = {
     [](struct wl_client*, struct wl_resource* surfaceResource)
     {
         auto& surface = *static_cast<Surface*>(wl_resource_get_user_data(surfaceResource));
-        if (!surface.exportableClient)
-            return;
-
-        struct wl_resource* bufferResource = surface.bufferResource;
-        surface.bufferResource = nullptr;
-
-        if (surface.dmabufBuffer)
-            surface.exportableClient->exportLinuxDmabuf(surface.dmabufBuffer);
-        else if (surface.shmBuffer)
-            surface.exportableClient->exportShmBuffer(bufferResource, surface.shmBuffer);
-        else
-            surface.exportableClient->exportBufferResource(bufferResource);
+        WS::Instance::singleton().impl().surfaceCommit(surface);
     },
     // set_buffer_transform
     [](struct wl_client*, struct wl_resource*, int32_t) { },
@@ -292,19 +216,26 @@ static const struct wpe_video_plane_display_dmabuf_interface s_wpeDmaBufInterfac
     },
 };
 
+static Instance* s_singleton;
+
+void Instance::construct(std::unique_ptr<Impl>&& impl)
+{
+    s_singleton = new Instance(std::move(impl));
+}
+
 Instance& Instance::singleton()
 {
-    static Instance* s_singleton;
-    if (!s_singleton)
-        s_singleton = new Instance;
+    assert(!!s_singleton);
     return *s_singleton;
 }
 
-Instance::Instance()
-    : m_display(wl_display_create())
+Instance::Instance(std::unique_ptr<Impl>&& impl)
+    : m_impl(std::move(impl))
+    , m_display(wl_display_create())
     , m_source(g_source_new(&ServerSource::s_sourceFuncs, sizeof(ServerSource)))
-    , m_eglDisplay(EGL_NO_DISPLAY)
 {
+    m_impl->setInstance(*this);
+
     m_compositor = wl_global_create(m_display, &wl_compositor_interface, 3, this,
         [](struct wl_client* client, void*, uint32_t version, uint32_t id)
         {
@@ -328,8 +259,6 @@ Instance::Instance()
             wl_resource_set_implementation(resource, &s_wpeBridgeInterface, nullptr, nullptr);
         });
 
-    wl_list_init(&m_dmabufBuffers);
-
     auto& source = *reinterpret_cast<ServerSource*>(m_source);
 
     struct wl_event_loop* eventLoop = wl_display_get_event_loop(m_display);
@@ -351,23 +280,13 @@ Instance::~Instance()
         g_source_unref(m_source);
     }
 
+    m_impl = nullptr;
+
     if (m_compositor)
         wl_global_destroy(m_compositor);
 
     if (m_wpeBridge)
         wl_global_destroy(m_wpeBridge);
-
-    if (m_linuxDmabuf) {
-        struct linux_dmabuf_buffer *buffer;
-        struct linux_dmabuf_buffer *tmp;
-        wl_list_for_each_safe(buffer, tmp, &m_dmabufBuffers, link) {
-            assert(buffer);
-
-            wl_list_remove(&buffer->link);
-            linux_dmabuf_buffer_destroy(buffer);
-        }
-        wl_global_destroy(m_linuxDmabuf);
-    }
 
     if (m_videoPlaneDisplayDmaBuf.object)
         wl_global_destroy(m_videoPlaneDisplayDmaBuf.object);
@@ -376,80 +295,9 @@ Instance::~Instance()
         wl_display_destroy(m_display);
 }
 
-static bool isEGLExtensionSupported(const char* extensionList, const char* extension)
-{
-    if (!extensionList)
-        return false;
-
-    int extensionLen = strlen(extension);
-    const char* extensionListPtr = extensionList;
-    while ((extensionListPtr = strstr(extensionListPtr, extension))) {
-        if (extensionListPtr[extensionLen] == ' ' || extensionListPtr[extensionLen] == '\0')
-            return true;
-	extensionListPtr += extensionLen;
-    }
-    return false;
-}
-
-bool Instance::initialize(EGLDisplay eglDisplay)
-{
-    if (m_eglDisplay == eglDisplay)
-        return true;
-
-    if (m_eglDisplay != EGL_NO_DISPLAY) {
-        g_warning("Multiple EGL displays are not supported.\n");
-        return false;
-    }
-
-    // wl_display_init_shm() returns `0` on success.
-    if (wl_display_init_shm(m_display) != 0)
-        return false;
-
-    const char* extensions = eglQueryString(eglDisplay, EGL_EXTENSIONS);
-    if (isEGLExtensionSupported(extensions, "EGL_WL_bind_wayland_display")) {
-        s_eglBindWaylandDisplayWL = reinterpret_cast<PFNEGLBINDWAYLANDDISPLAYWL>(eglGetProcAddress("eglBindWaylandDisplayWL"));
-        assert(s_eglBindWaylandDisplayWL);
-        s_eglQueryWaylandBufferWL = reinterpret_cast<PFNEGLQUERYWAYLANDBUFFERWL>(eglGetProcAddress("eglQueryWaylandBufferWL"));
-        assert(s_eglQueryWaylandBufferWL);
-    }
-
-    if (isEGLExtensionSupported(extensions, "EGL_KHR_image_base")) {
-        s_eglCreateImageKHR = reinterpret_cast<PFNEGLCREATEIMAGEKHRPROC>(eglGetProcAddress("eglCreateImageKHR"));
-        assert(s_eglCreateImageKHR);
-        s_eglDestroyImageKHR = reinterpret_cast<PFNEGLDESTROYIMAGEKHRPROC>(eglGetProcAddress("eglDestroyImageKHR"));
-        assert(s_eglDestroyImageKHR);
-    }
-
-    if (s_eglBindWaylandDisplayWL && s_eglQueryWaylandBufferWL) {
-        if (!s_eglCreateImageKHR || !s_eglDestroyImageKHR)
-            return false;
-        if (!s_eglBindWaylandDisplayWL(eglDisplay, m_display))
-            return false;
-    }
-
-    m_eglDisplay = eglDisplay;
-
-    /* Initialize Linux dmabuf subsystem. */
-    if (isEGLExtensionSupported(extensions, "EGL_EXT_image_dma_buf_import")
-        && isEGLExtensionSupported(extensions, "EGL_EXT_image_dma_buf_import_modifiers")) {
-        s_eglQueryDmaBufFormatsEXT = reinterpret_cast<PFNEGLQUERYDMABUFFORMATSEXTPROC>(eglGetProcAddress("eglQueryDmaBufFormatsEXT"));
-        assert(s_eglQueryDmaBufFormatsEXT);
-        s_eglQueryDmaBufModifiersEXT = reinterpret_cast<PFNEGLQUERYDMABUFMODIFIERSEXTPROC>(eglGetProcAddress("eglQueryDmaBufModifiersEXT"));
-        assert(s_eglQueryDmaBufModifiersEXT);
-    }
-
-    if (s_eglQueryDmaBufFormatsEXT && s_eglQueryDmaBufModifiersEXT) {
-        if (m_linuxDmabuf)
-            assert(!"Linux-dmabuf has already been initialized");
-        m_linuxDmabuf = linux_dmabuf_setup(m_display);
-    }
-
-    return true;
-}
-
 int Instance::createClient()
 {
-    if (m_eglDisplay == EGL_NO_DISPLAY)
+    if (!m_impl->initialized())
         return -1;
 
     int pair[2];
@@ -466,149 +314,6 @@ int Instance::createClient()
 void Instance::registerSurface(uint32_t id, Surface* surface)
 {
     m_viewBackendMap.insert({ id, surface });
-}
-
-EGLImageKHR Instance::createImage(struct wl_resource* resourceBuffer)
-{
-    if (m_eglDisplay == EGL_NO_DISPLAY)
-        return EGL_NO_IMAGE_KHR;
-    return s_eglCreateImageKHR(m_eglDisplay, EGL_NO_CONTEXT, EGL_WAYLAND_BUFFER_WL, resourceBuffer, nullptr);
-}
-
-EGLImageKHR Instance::createImage(const struct linux_dmabuf_buffer* dmabufBuffer)
-{
-    if (!m_linuxDmabuf)
-        return EGL_NO_IMAGE_KHR;
-
-    static const struct {
-        EGLint fd;
-        EGLint offset;
-        EGLint pitch;
-        EGLint modifierLo;
-        EGLint modifierHi;
-    } planeEnums[4] = {
-        {EGL_DMA_BUF_PLANE0_FD_EXT,
-         EGL_DMA_BUF_PLANE0_OFFSET_EXT,
-         EGL_DMA_BUF_PLANE0_PITCH_EXT,
-         EGL_DMA_BUF_PLANE0_MODIFIER_LO_EXT,
-         EGL_DMA_BUF_PLANE0_MODIFIER_HI_EXT},
-        {EGL_DMA_BUF_PLANE1_FD_EXT,
-         EGL_DMA_BUF_PLANE1_OFFSET_EXT,
-         EGL_DMA_BUF_PLANE1_PITCH_EXT,
-         EGL_DMA_BUF_PLANE1_MODIFIER_LO_EXT,
-         EGL_DMA_BUF_PLANE1_MODIFIER_HI_EXT},
-        {EGL_DMA_BUF_PLANE2_FD_EXT,
-         EGL_DMA_BUF_PLANE2_OFFSET_EXT,
-         EGL_DMA_BUF_PLANE2_PITCH_EXT,
-         EGL_DMA_BUF_PLANE2_MODIFIER_LO_EXT,
-         EGL_DMA_BUF_PLANE2_MODIFIER_HI_EXT},
-        {EGL_DMA_BUF_PLANE3_FD_EXT,
-         EGL_DMA_BUF_PLANE3_OFFSET_EXT,
-         EGL_DMA_BUF_PLANE3_PITCH_EXT,
-         EGL_DMA_BUF_PLANE3_MODIFIER_LO_EXT,
-         EGL_DMA_BUF_PLANE3_MODIFIER_HI_EXT},
-    };
-
-    EGLint attribs[50];
-    int atti = 0;
-    attribs[atti++] = EGL_WIDTH;
-    attribs[atti++] = dmabufBuffer->attributes.width;
-    attribs[atti++] = EGL_HEIGHT;
-    attribs[atti++] = dmabufBuffer->attributes.height;
-    attribs[atti++] = EGL_LINUX_DRM_FOURCC_EXT;
-    attribs[atti++] = dmabufBuffer->attributes.format;
-
-    for (int i = 0; i < dmabufBuffer->attributes.n_planes; i++) {
-        attribs[atti++] = planeEnums[i].fd;
-        attribs[atti++] = dmabufBuffer->attributes.fd[i];
-        attribs[atti++] = planeEnums[i].offset;
-        attribs[atti++] = dmabufBuffer->attributes.offset[i];
-        attribs[atti++] = planeEnums[i].pitch;
-        attribs[atti++] = dmabufBuffer->attributes.stride[i];
-        attribs[atti++] = planeEnums[i].modifierLo;
-        attribs[atti++] = dmabufBuffer->attributes.modifier[i] & 0xFFFFFFFF;
-        attribs[atti++] = planeEnums[i].modifierHi;
-        attribs[atti++] = dmabufBuffer->attributes.modifier[i] >> 32;
-    }
-
-    attribs[atti++] = EGL_NONE;
-
-    return s_eglCreateImageKHR(m_eglDisplay, EGL_NO_CONTEXT, EGL_LINUX_DMA_BUF_EXT, nullptr, attribs);
-}
-
-void Instance::destroyImage(EGLImageKHR image)
-{
-    if (m_eglDisplay == EGL_NO_DISPLAY)
-        return;
-    s_eglDestroyImageKHR(m_eglDisplay, image);
-}
-
-void Instance::queryBufferSize(struct wl_resource* bufferResource, uint32_t* width, uint32_t* height)
-{
-    if (width) {
-        int w;
-        s_eglQueryWaylandBufferWL(m_eglDisplay, bufferResource, EGL_WIDTH, &w);
-        *width = w;
-    }
-
-    if (height) {
-        int h;
-        s_eglQueryWaylandBufferWL(m_eglDisplay, bufferResource, EGL_HEIGHT, &h);
-        *height = h;
-    }
-}
-
-void Instance::importDmaBufBuffer(struct linux_dmabuf_buffer* buffer)
-{
-    if (!m_linuxDmabuf)
-        return;
-    wl_list_insert(&m_dmabufBuffers, &buffer->link);
-}
-
-const struct linux_dmabuf_buffer* Instance::getDmaBufBuffer(struct wl_resource* bufferResource) const
-{
-    if (!m_linuxDmabuf || !bufferResource)
-        return nullptr;
-
-    if (!linux_dmabuf_buffer_implements_resource(bufferResource))
-        return nullptr;
-
-    struct linux_dmabuf_buffer* buffer;
-    wl_list_for_each(buffer, &m_dmabufBuffers, link) {
-        if (buffer->buffer_resource == bufferResource)
-            return buffer;
-    }
-
-    return NULL;
-}
-
-void Instance::foreachDmaBufModifier(std::function<void (int format, uint64_t modifier)> callback)
-{
-    if (m_eglDisplay == EGL_NO_DISPLAY)
-        return;
-
-    EGLint formats[128];
-    EGLint numFormats;
-    if (!s_eglQueryDmaBufFormatsEXT(m_eglDisplay, 128, formats, &numFormats))
-        assert(!"Linux-dmabuf: Failed to query formats");
-
-    for (int i = 0; i < numFormats; i++) {
-        uint64_t modifiers[64];
-        EGLint numModifiers;
-        if (!s_eglQueryDmaBufModifiersEXT(m_eglDisplay, formats[i], 64, modifiers, NULL, &numModifiers))
-            assert(!"Linux-dmabuf: Failed to query modifiers of a format");
-
-        /* Send DRM_FORMAT_MOD_INVALID token when no modifiers are supported
-         * for this format.
-         */
-        if (numModifiers == 0) {
-            numModifiers = 1;
-            modifiers[0] = DRM_FORMAT_MOD_INVALID;
-        }
-
-        for (int j = 0; j < numModifiers; j++)
-            callback(formats[i], modifiers[j]);
-    }
 }
 
 void Instance::initializeVideoPlaneDisplayDmaBuf(VideoPlaneDisplayDmaBufCallback updateCallback, VideoPlaneDisplayDmaBufEndOfStreamCallback endOfStreamCallback)
